Split DVCClient tests into multiple files with common DVCClientTestBase class
